Moved Transaction extension to its own file
