+++ Sources/SuperwallKit/StoreKit/Transactions/Purchasing/PurchasingCoordinator.swift
@@ -8,25 +8,6 @@
 import Foundation
 import StoreKit
 
-@available(iOS 15.0, *)
-extension Transaction {
-  /// Gets the latest transaction for a given `productId` since
-  /// an hour up to a given `purchaseDate`.
-  static func latest(
-    for productId: String,
-    since purchaseDate: Date
-  ) async -> VerificationResult<Transaction>? {
-    let verificationResult = await Transaction.latest(for: productId)
-
-    if let transaction = verificationResult.map({ $0.unsafePayloadValue }),
-      purchaseDate.addingTimeInterval(-3600).compare(transaction.purchaseDate) == .orderedAscending {
-      return verificationResult
-    }
-
-    return nil
-  }
-}
-
 /// An actor that manages and coordinates the storing of types associated with purchasing.
 actor PurchasingCoordinator {
   private var completion: ((PurchaseResult) -> Void)?
